requirements.txt
pandas
apify-client
google-generativeai
python-dotenv
streamlit
plotly